Move beans to IWXXMGenericAviationWeatherMessageConverter
