+++ src/main/java/fi/fmi/avi/converter/iwxxm/bulletin/v1_2/IWXXMGenericBulletinScanner.java
@@ -1,10 +1,9 @@
-package fi.fmi.avi.converter.iwxxm.generic;
+package fi.fmi.avi.converter.iwxxm.bulletin.v1_2;
 
 import org.w3c.dom.Element;
 
 import fi.fmi.avi.converter.ConversionHints;
 import fi.fmi.avi.converter.ConversionResult;
-import fi.fmi.avi.converter.iwxxm.bulletin.v1_2.MeteorologicalBulletinIWXXMScanner;
 import fi.fmi.avi.model.GenericAviationWeatherMessage;
 import fi.fmi.avi.model.bulletin.GenericMeteorologicalBulletin;
